refactor: split screens under detekt's per-file function/line budgets; fix ktlint line lengths (tc-oibm)
